requirements.txt
openai
google-cloud-firestore
elevenlabs
streamlit
python-dotenv
#sudo apt update
#sudo apt install ffmpeg
# install mpv through Software GUI in Ubuntu